# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  api.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/api.APIError'
      meta:
        $ref: '#/definitions/api.Meta'
      success:
        type: boolean
    type: object
  api.Meta:
    properties:
      hidden_unresolved_telegram_count:
        type: integer
      pagination:
        $ref: '#/definitions/api.PaginationMeta'
    type: object
  api.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
...
  handlers.WhatsAppBackfillResponse:
    properties:
      dropped_inline_chunks:
        description: |-
          DroppedInlineChunks counts bootstrap chunks the server delivered with
          history embedded inline against our explicit non-inline request. Those
          are dropped un-projected rather than persisted, so a non-zero count is a
          visible, accepted gap in backfill.
        type: integer
      failed:
        type: integer
      observed_floor_at:
        format: date-time
        type: string
      pending:
        type: integer
      processing:
        type: integer
    type: object
  handlers.WhatsAppDisconnectResponse:
    properties:
      already_unlinked:
        description: |-
          AlreadyUnlinked is true when the device was confirmed gone server-side
          before the call, so no remote unlink was attempted.
        type: boolean
      forced:
        description: |-
          Forced is true when local credentials were cleared without server
          confirmation.
        type: boolean
      remote_unlinked:
        description: RemoteUnlinked is true when this call logged the device out server-side.
        type: boolean
      warning:
        description: Warning is set when the user must finish the unlink from their
          phone.
        type: string
    type: object
  handlers.WhatsAppPairRequest:
    properties:
      method:
        description: Method is "qr" or "phone".
        enum:
        - qr
        - phone
        type: string
      phone:
        description: Phone is required when Method is "phone", in E.164 form (+15551234567).
        type: string
    required:
    - method
    type: object
  handlers.WhatsAppPairingResponse:
    properties:
      expires_at:
        format: date-time
        type: string
      method:
        type: string
      pair_code:
        description: PairCode is the 8-character code the user types on their phone.
        type: string
      qr_code:
        description: QRCode is the raw whatsmeow code string; the client renders the
          image.
        type: string
    type: object
  handlers.WhatsAppStatusResponse:
    properties:
      backfill:
        $ref: '#/definitions/handlers.WhatsAppBackfillResponse'
      banned_until:
        format: date-time
        type: string
      configured:
        type: boolean
      connected_at:
        format: date-time
        type: string
      jid:
        type: string
      pairing:
        $ref: '#/definitions/handlers.WhatsAppPairingResponse'
      phone_number:
        type: string
      push_name:
        type: string
      reason:
        description: |-
          Reason carries the machine-readable detail for not_ready, disconnected,
          disconnect_failed and error.
        type: string
      state:
        description: |-
          State is one of not_ready, not_paired, pairing, connecting, connected,
          reconnecting, disconnected, disconnect_failed, error.
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A personal customer relationship management API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Personal CRM API
  version: "1.0"
paths:
...
  /whatsapp/auth:
    delete:
      description: Unlink the WhatsApp device. Local credentials are kept when the
        remote unlink fails, so the user can retry; ?force=true clears them locally
        without contacting WhatsApp and returns a warning that the device must be
        unlinked from the phone.
      parameters:
      - description: Clear local credentials without contacting WhatsApp
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.WhatsAppDisconnectResponse'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
        "502":
          description: Bad Gateway
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
      summary: Disconnect WhatsApp
      tags:
      - whatsapp
  /whatsapp/auth/cancel:
    post:
      description: Abort an in-flight WhatsApp pairing attempt. Idempotent.
      produces:
      - application/json
      responses:
        "204":
          description: Pairing cancelled
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
      summary: Cancel WhatsApp pairing
      tags:
      - whatsapp
  /whatsapp/auth/start:
    post:
      consumes:
      - application/json
      description: Begin linking a WhatsApp account by QR code or phone pairing code.
        The first code is returned in the status body; poll GET /whatsapp/auth/status
        for refreshed codes.
      parameters:
      - description: Pairing request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.WhatsAppPairRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.WhatsAppStatusResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
        "504":
          description: Gateway Timeout
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
      summary: Start WhatsApp pairing
      tags:
      - whatsapp
  /whatsapp/auth/status:
    get:
      description: Report the WhatsApp connection state, the linked account when paired,
        any in-flight pairing code, and the history-backfill counts. Absent (404)
        when the feature is disabled.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.WhatsAppStatusResponse'
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/api.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/api.APIError'
              type: object
      summary: Get WhatsApp status
      tags:
      - whatsapp
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"